Move subclauses
