+++ packages/graphql/src/translate/cypher-builder/CypherBuilder.ts
@@ -101,7 +101,7 @@ export type { BooleanOp } from "./operations/boolean";
 export type { Expr, Predicate } from "./types";
 export type { CypherFunction as Function } from "./functions/CypherFunction";
 export type { ComprehensionExpr } from "./expressions/list/ComprehensionExpr";
-export type { ProjectionColumn } from "./sub-clauses/Projection";
-export type { SetParam } from "./sub-clauses/Set";
+export type { ProjectionColumn } from "./clauses/sub-clauses/Projection";
+export type { SetParam } from "./clauses/sub-clauses/Set";
 export type { PredicateFunction } from "./functions/PredicateFunctions";
-export type { Order } from "./sub-clauses/OrderBy";
+export type { Order } from "./clauses/sub-clauses/OrderBy";
