+++ packages/graphql/src/translate/cypher-builder/clauses/Call.ts
@@ -22,7 +22,7 @@ import type { CypherASTNode } from "../CypherASTNode";
 import type { Variable } from "../variables/Variable";
 import { Clause } from "./Clause";
 import { compileCypherIfExists, padBlock } from "../utils/utils";
-import { ImportWith } from "../sub-clauses/ImportWith";
+import { ImportWith } from "./sub-clauses/ImportWith";
 import { applyMixins } from "./utils/apply-mixin";
 import { WithReturn } from "./mixins/WithReturn";
