+++ packages/graphql/src/translate/cypher-builder/clauses/Match.ts
@@ -26,9 +26,9 @@ import { applyMixins } from "./utils/apply-mixin";
 import { WithWhere } from "./mixins/WithWhere";
 import { WithSet } from "./mixins/WithSet";
 import { WithWith } from "./mixins/WithWith";
-import { DeleteClause, DeleteInput } from "../sub-clauses/Delete";
+import { DeleteClause, DeleteInput } from "./sub-clauses/Delete";
 import type { PropertyRef } from "../PropertyRef";
-import { RemoveClause } from "../sub-clauses/Remove";
+import { RemoveClause } from "./sub-clauses/Remove";
 
 export class Match<T extends MatchableElement = any> extends Clause {
     private pattern: Pattern<T>;
