+++ packages/graphql/src/translate/cypher-builder/clauses/Merge.ts
@@ -22,7 +22,7 @@ import type { RelationshipRef } from "../variables/RelationshipRef";
 import { NodeRef } from "../variables/NodeRef";
 import { MatchParams, Pattern } from "../Pattern";
 import { Clause } from "./Clause";
-import { OnCreate, OnCreateParam } from "../sub-clauses/OnCreate";
+import { OnCreate, OnCreateParam } from "./sub-clauses/OnCreate";
 import { WithReturn } from "./mixins/WithReturn";
 import { applyMixins } from "./utils/apply-mixin";
