+++ packages/graphql/src/translate/cypher-builder/clauses/mixins/WithReturn.ts
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-import type { ProjectionColumn } from "../../sub-clauses/Projection";
+import type { ProjectionColumn } from "../sub-clauses/Projection";
 import { Return } from "../Return";
 import { ClauseMixin } from "./ClauseMixin";
