+++ packages/graphql/src/translate/cypher-builder/clauses/mixins/WithSet.ts
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-import { SetClause, SetParam } from "../../sub-clauses/Set";
+import { SetClause, SetParam } from "../sub-clauses/Set";
 import { ClauseMixin } from "./ClauseMixin";
 
 export abstract class WithSet extends ClauseMixin {
