+++ packages/graphql/src/translate/cypher-builder/clauses/sub-clauses/OrderBy.ts
@@ -17,11 +17,11 @@
  * limitations under the License.
  */
 
-import type { CypherEnvironment } from "../Environment";
+import type { CypherEnvironment } from "../../Environment";
 
-import { CypherASTNode } from "../CypherASTNode";
-import type { Expr } from "../types";
-import { compileCypherIfExists } from "../utils/utils";
+import { CypherASTNode } from "../../CypherASTNode";
+import type { Expr } from "../../types";
+import { compileCypherIfExists } from "../../utils/utils";
 
 export type Order = "ASC" | "DESC";
