+++ packages/graphql/src/translate/cypher-builder/clauses/sub-clauses/Projection.ts
@@ -17,10 +17,10 @@
  * limitations under the License.
  */
 
-import type { CypherEnvironment } from "../Environment";
-import type { Expr } from "../types";
-import type { Literal, Variable } from "../CypherBuilder";
-import { CypherASTNode } from "../CypherASTNode";
+import type { CypherEnvironment } from "../../Environment";
+import type { Expr } from "../../types";
+import type { Literal, Variable } from "../../CypherBuilder";
+import { CypherASTNode } from "../../CypherASTNode";
 
 // TODO: improve projection column, some expressions require alias
 export type ProjectionColumn = Expr | [Expr, string | Variable | Literal];
