+++ packages/graphql/src/translate/cypher-builder/clauses/sub-clauses/Set.ts
@@ -17,11 +17,11 @@
  * limitations under the License.
  */
 
-import type { CypherASTNode } from "../CypherASTNode";
-import type { CypherEnvironment } from "../Environment";
-import type { PropertyRef } from "../PropertyRef";
-import type { Expr } from "../types";
-import { padBlock } from "../utils/utils";
+import type { CypherASTNode } from "../../CypherASTNode";
+import type { CypherEnvironment } from "../../Environment";
+import type { PropertyRef } from "../../PropertyRef";
+import type { Expr } from "../../types";
+import { padBlock } from "../../utils/utils";
 import { SubClause } from "./SubClause";
 
 export type SetParam = [PropertyRef, Expr];
